1. Весь код не написан в одном классе Main. 2. Повторяющийся блок заменен циклом. 3. При вводе данных сразу вычисляется победитель. 4. Константы для максимальной и минимальной скорости. 5. Константы для полей CAR_NAME и CAR_SPEED.
