+++ src/main/java/Car.java
@@ -1,0 +1,10 @@
+// Класс «Автомобиль»
+public class Car {
+    final String CAR_NAME;
+    final int CAR_SPEED;
+
+    public Car(String carName, int carSpeed) {
+        this.CAR_NAME = carName;
+        this.CAR_SPEED = carSpeed;
+    }
+}
